% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pointdensity.R
\name{addCheckToGeom}
\alias{addCheckToGeom}
\title{Wraps the user supplied Geom (typically GeomPoint) to add class "check_aspect_grob" and information about the aspect ratio assumed under which the densities were calculated to the grobs it draws
The check is injected by providing an S3 instance to the makeContext generic that is called by grid for each grob at render time (where the actual plot aspect ratio is finally known)}
\usage{
addCheckToGeom(orig_geom, expected_aspect_ratio = 1)
}
\description{
Wraps the user supplied Geom (typically GeomPoint) to add class "check_aspect_grob" and information about the aspect ratio assumed under which the densities were calculated to the grobs it draws
The check is injected by providing an S3 instance to the makeContext generic that is called by grid for each grob at render time (where the actual plot aspect ratio is finally known)
}
