% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pointdensity.R
\name{count_neighbors_r}
\alias{count_neighbors_r}
\title{Implementation of count_neighbors in R. Not actually used, just for clarity}
\usage{
count_neighbors_r(x, y, r2, xy)
}
\description{
Implementation of count_neighbors in R. Not actually used, just for clarity
}
