% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpointdensity-package.R
\docType{package}
\name{ggpointdensity-package}
\alias{ggpointdensity}
\alias{ggpointdensity-package}
\title{ggpointdensity: A Cross Between a 2D Density Plot and a Scatter Plot}
\description{
A cross between a 2D density plot and a scatter plot, implemented as a 'ggplot2' geom. Points in the scatter plot are colored by the number of neighboring points. This is useful to visualize the 2D-distribution of points in case of overplotting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LKremer/ggpointdensity}
  \item Report bugs at \url{https://github.com/LKremer/ggpointdensity/issues}
}

}
\author{
\strong{Maintainer}: Lukas P. M. Kremer \email{L-Kremer@web.de} (\href{https://orcid.org/0000-0003-3170-6295}{ORCID})

Other contributors:
\itemize{
  \item Simon Anders (\href{https://orcid.org/0000-0003-4868-1805}{ORCID}) [contributor]
}

}
\keyword{internal}
