% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pointdensity.R
\docType{data}
\name{StatPointdensity}
\alias{StatPointdensity}
\title{ggproto class, see \code{\link[ggplot2:Stat]{ggplot2::Stat()}}}
\description{
ggproto class, see \code{\link[ggplot2:Stat]{ggplot2::Stat()}}
}
\keyword{datasets}
